// sdk/bindings/AppToServer.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { InitializeRequest } from "./InitializeRequest";
import type { RequestPayload } from "./RequestPayload";

export type AppToServer = { type: "InitializeRequest" } & InitializeRequest | { type: "RequestPayload" } & RequestPayload;